# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Reservas Travelamedida",
        "description": "Documentación de la API para el sistema de reservas Travelamedida",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "alexvazu@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/auth",
            "description": "Servidor API Principal"
        }
    ],
    "paths": {
...
        "/pagos/crear-pendiente": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Crear un pago pendiente",
                "operationId": "9f8aae41dc5a80c56f5767755e066628",
                "parameters": [
                    {
                        "name": "X-Reserva-Segura",
                        "in": "header",
                        "description": "Cabecera de seguridad privada para uso interno",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "clave-super-secreta-123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "expiration_token",
                                    "usuario_id",
                                    "monto",
                                    "metodo_pago"
                                ],
                                "properties": {
                                    "guid": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "obtener-del-metodo-generar-guid"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "AQUI-VA-EL-TOKEN-DE-VISION-MUNDO"
                                    },
                                    "expiration_token": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1650326400
                                    },
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "transaction_id_metodo_pago": {
                                        "type": "string",
                                        "example": "tx_123456",
                                        "nullable": true
                                    },
                                    "metodo_pago": {
                                        "type": "string",
                                        "enum": [
                                            "StereumPay",
                                            "tarjeta"
                                        ],
                                        "example": "StereumPay"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pago pendiente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "guid": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "expiration_token": {
                                            "type": "integer"
                                        },
                                        "usuario_id": {
                                            "type": "integer"
                                        },
                                        "monto": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "transaction_id_metodo_pago": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "metodo_pago": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "token": [
                                                    "The token field is required."
                                                ],
                                                "monto": [
                                                    "El monto debe ser al menos 2.01."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pagos/generar-guid": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Generar un GUID único para pagos",
                "operationId": "db0fd9ac250a36370bb2a34cb41691f1",
                "parameters": [
                    {
                        "name": "X-Reserva-Segura",
                        "in": "header",
                        "description": "Cabecera de seguridad privada para uso interno",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "clave-super-secreta-123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GUID generado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "guid": {
                                            "type": "string",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token, formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}